+++ frontend/src/features/map_planning/routes/MapWrapper.tsx
@@ -9,6 +9,7 @@ import { useSafeAuth } from '@/hooks/useSafeAuth';
 import { EditorMap } from '../components/EditorMap';
 import {
   useBaseLayer,
+  useDrawingLayer,
   useGetLayers,
   useInvalidateMapQueries,
   useMap,
@@ -62,8 +63,13 @@ const useInitializeTimeline = (mapId: number) => {
 function useInitializeMap() {
   const mapId = useMapId();
   const { data: map } = useMap(mapId);
-  const { t } = useTranslation(['layers']);
   const { data: layers, error } = useGetLayers(mapId);
+  const { t } = useTranslation(['layers']);
+
+  if (error) {
+    console.error(error);
+    errorToastGrouped(t('layers:error_fetching_layers'), { autoClose: false });
+  }
 
   if (error) {
     console.error(error);
@@ -82,12 +88,19 @@ function useInitializeMap() {
 
   const plantLayer = getDefaultLayer(LayerType.Plants, layers);
   const baseLayer = getDefaultLayer(LayerType.Base, layers);
+  const drawingLayer = getDefaultLayer(LayerType.Drawing, layers);
 
   // The casts are fine because we know that the queries only execute once they are enabled.
   usePlantLayer({
     mapId,
     layerId: plantLayer?.id as number,
     enabled: Boolean(plantLayer),
+  });
+
+  useDrawingLayer({
+    mapId,
+    layerId: drawingLayer?.id as number,
+    enabled: Boolean(drawingLayer),
   });
 
   useBaseLayer({