+++ frontend/src/features/map_planning/utils/layer-utils.ts
@@ -1,3 +1,4 @@
+import { LayerType } from '@/api_types/definitions';
 import { FrontendOnlyLayerType } from '../layers/_frontend_only';
 import useMapStore from '../store/MapStore';
 import { CombinedLayerType } from '../store/MapStoreTypes';
@@ -14,10 +15,15 @@ export function useSelectedLayerType(): CombinedLayerType {
 
 export function useIsPlantLayerActive(): boolean {
   const selectedLayerType = useSelectedLayerType();
-  return selectedLayerType === 'plants';
+  return selectedLayerType === LayerType.Plants;
 }
 
 export function useIsBaseLayerActive(): boolean {
   const selectedLayerType = useSelectedLayerType();
-  return selectedLayerType === 'base';
+  return selectedLayerType === LayerType.Base;
+}
+
+export function useIsDrawingLayerActive(): boolean {
+  const selectedLayerType = useSelectedLayerType();
+  return selectedLayerType === LayerType.Drawing;
 }